requirements.txt
plexapi
tmdbv3api
requests
pyyaml
flask
python-dotenv